Refactor command imports to use the updated Flag class and remove redundant Flag class
