+++ src/main/java/modmate/command/ViewModCommand.java
@@ -2,6 +2,7 @@ package modmate.command;
 
 import java.util.List;
 
+import modmate.command.util.Flag;
 import modmate.download.nusmods.NUSModsAPI;
 import modmate.exception.ApiException;
 import modmate.exception.CommandException;
